+++ .github/workflows/test.yml
@@ -51,7 +51,7 @@ jobs:
           BITTENSOR_WALLET_HOTKEY_NAME: test_hotkey
           SQLALCHEMY_DATABASE_URI: sqlite:///test.db
           ASYNC_SQLALCHEMY_DATABASE_URI: sqlite+aiosqlite:///test.db
-        run: pdm run pytest tests/ -v --tb=short
+        run: pdm run pytest tests/ -v --tb=short --strict-markers
 
   # Executor tests - no external services needed
   executor:
@@ -82,7 +82,7 @@ jobs:
         run: pdm install
 
       - name: Run tests
-        run: pdm run pytest tests/ -v --tb=short
+        run: pdm run pytest tests/ -v --tb=short --strict-markers
 
   # Miners tests - requires PostgreSQL
   miners:
@@ -133,4 +133,4 @@ jobs:
           BITTENSOR_WALLET_HOTKEY_NAME: test_hotkey
           SQLALCHEMY_DATABASE_URI: postgresql://postgres:password@localhost:5433/miner_test
           EXTERNAL_IP_ADDRESS: 127.0.0.1
-        run: pdm run pytest tests/ -v --tb=short
+        run: pdm run pytest tests/ -v --tb=short --strict-markers
