+++ neurons/validators/tests/conftest.py
@@ -1,7 +1,16 @@
+"""Pytest configuration and fixtures for validators tests.
+
+This module provides shared fixtures for all test modules.
+Fixtures are organized into separate modules for better maintainability:
+- fixtures/db_fixtures.py: Database-related fixtures
+- fixtures/mock_fixtures.py: Mock objects and services
+- helpers/context.py: Context builders for pipeline tests
+- factories/port_mapping.py: Factory functions for test data
+"""
+
 import logging
 import sys
 from pathlib import Path
-from unittest.mock import AsyncMock, MagicMock, patch
 
 # Add validators directory to path FIRST (before lium-io root which also has tests/)
 VALIDATORS_ROOT = Path(__file__).resolve().parents[1]
@@ -15,25 +24,37 @@ if lium_root not in sys.path:
     sys.path.append(lium_root)
 
 import pytest
-import pytest_asyncio
-from datura.requests.miner_requests import ExecutorSSHInfo
-from sqlalchemy.ext.asyncio import create_async_engine
-from sqlmodel import SQLModel
-from sqlmodel.ext.asyncio.session import AsyncSession
 
-from tests.helpers import make_context
+# Import fixtures from organized modules
+from tests.fixtures.db_fixtures import test_db_session, test_engine  # noqa: E402
+from tests.fixtures.mock_fixtures import (  # noqa: E402
+    mock_aiohttp_session,
+    mock_async_session_maker,
+    mock_redis_service,
+    mock_ssh_client,
+    port_mapping_dao,
+    sample_executor_info,
+)
+from tests.helpers.context import make_context  # noqa: E402
 
 
 # Legacy fixture for pipeline tests
 @pytest.fixture
 def context_factory():
+    """Factory fixture for creating pipeline Context objects.
+
+    Usage:
+        def test_something(context_factory):
+            ctx = context_factory(miner_hotkey="test-key")
+            # ... use ctx
+    """
+
     def _factory(**overrides):
         return make_context(**overrides)
 
     return _factory
 
 
-# Fixtures from src/tests for new service tests
 @pytest.fixture(scope="session", autouse=True)
 def setup_sql_logging():
     """Enable SQL query logging for all tests."""
@@ -48,113 +69,16 @@ def setup_sql_logging():
     print("SQL logging enabled for all tests")
 
 
-@pytest.fixture
-def mock_ssh_client():
-    """Mock SSH client for testing Docker operations."""
-    client = AsyncMock()
-    client.run = AsyncMock(return_value=MagicMock(exit_status=0, stdout="", stderr=""))
-    return client
-
-
-@pytest.fixture
-def mock_redis_service():
-    """Mock Redis service for testing port storage operations."""
-    service = AsyncMock()
-    service.lpush = AsyncMock()
-    service.lrem = AsyncMock()
-    service.lrange = AsyncMock(return_value=[])
-    service.rpop = AsyncMock()
-    return service
-
-
-@pytest.fixture
-def sample_executor_info():
-    """Sample ExecutorSSHInfo for testing."""
-    port_mappings = [[9000 + i, 9000 + i] for i in range(1005)]
-    return ExecutorSSHInfo(
-        uuid="a1b2c3d4-e5f6-7890-abcd-ef1234567890",
-        address="192.168.1.100",
-        port=8080,
-        ssh_username="root",
-        ssh_port=22,
-        port_mappings=str(port_mappings),
-        port_range="40000-50000",
-        python_path="/usr/bin/python3",
-        root_dir="/tmp",
-    )
-
-
-@pytest.fixture
-def mock_aiohttp_session():
-    """Mock aiohttp session for testing HTTP requests."""
-    with patch("aiohttp.ClientSession") as mock_session_class:
-        session = AsyncMock()
-        mock_session_class.return_value.__aenter__.return_value = session
-
-        response = AsyncMock()
-        response.status = 200
-        response.json = AsyncMock(return_value={"status": "ok"})
-        session.get.return_value.__aenter__.return_value = response
-        session.post.return_value.__aenter__.return_value = response
-
-        yield session
-
-
-@pytest_asyncio.fixture(scope="session")
-async def test_engine():
-    """Create an async engine for testing with SQLite in-memory database."""
-    engine = create_async_engine(
-        "sqlite+aiosqlite:///:memory:",
-        echo=True,
-        future=True,
-    )
-
-    async with engine.begin() as conn:
-        await conn.run_sync(SQLModel.metadata.create_all)
-
-    yield engine
-
-    await engine.dispose()
-
-
-@pytest_asyncio.fixture
-async def test_db_session(test_engine):
-    """Create a test database session for each test."""
-    from sqlalchemy.orm import sessionmaker
-
-    async_session_maker = sessionmaker(
-        bind=test_engine,
-        class_=AsyncSession,
-        expire_on_commit=False,
-    )
-
-    async with async_session_maker() as session:
-        yield session
-        await session.rollback()
-
-
-@pytest.fixture
-def mock_async_session_maker(test_db_session):
-    """Mock the global AsyncSessionMaker to use test database."""
-
-    class MockContextManager:
-        def __init__(self, session):
-            self._session = session
-
-        async def __aenter__(self):
-            return self._session
-
-        async def __aexit__(self, exc_type, exc_val, exc_tb):
-            if exc_type:
-                await self._session.rollback()
-
-    with patch('daos.base.AsyncSessionMaker') as mock_maker:
-        mock_maker.return_value = MockContextManager(test_db_session)
-        yield mock_maker
-
-
-@pytest.fixture
-def port_mapping_dao(mock_async_session_maker):
-    """Create PortMappingDao for testing with test database."""
-    from daos.port_mapping_dao import PortMappingDao
-    return PortMappingDao()
+# Re-export all fixtures for pytest discovery
+__all__ = [
+    "context_factory",
+    "mock_aiohttp_session",
+    "mock_async_session_maker",
+    "mock_redis_service",
+    "mock_ssh_client",
+    "port_mapping_dao",
+    "sample_executor_info",
+    "setup_sql_logging",
+    "test_db_session",
+    "test_engine",
+]
