+++ neurons/validators/tests/daos/__init__.py
@@ -1,0 +1,1 @@
+"""DAO tests."""
