+++ neurons/validators/tests/factories/__init__.py
@@ -1,0 +1,11 @@
+"""Test factories for validators tests."""
+
+from tests.factories.port_mapping import (
+    create_port_mapping,
+    create_port_mappings_batch,
+)
+
+__all__ = [
+    "create_port_mapping",
+    "create_port_mappings_batch",
+]
